稳定版本

+++ OrbWalker/object.h
@@ -7,7 +7,7 @@
 class Object {
 public:
 	union {
-		DEFINE_MEMBER_N(int, team, 0x34);
+		DEFINE_MEMBER_N(short, team, 0x34);
 		DEFINE_MEMBER_N(Vector3, position, 0x1DC);
 		DEFINE_MEMBER_N(bool, visible, 0x274); 
 		DEFINE_MEMBER_N(bool, targetable, 0xD04);
